requirements.txt
flask==3.1.1
requests==2.19.1

argon2-cffi==25.1.0
